+++ webviz_subsurface/plugins/_well_analysis/_types.py
@@ -16,3 +16,13 @@ class ChartType(StrEnum):
     BAR = "bar"
     PIE = "pie"
     AREA = "area"
+
+
+class StatType(StrEnum):
+    MEAN = "mean"
+    P10 = "p10"
+    P50 = "p50"
+    P90 = "p90"
+    MAX = "max"
+    MIN = "min"
+    P10_MINUS_P90 = "p10-p90"