+++ webviz_subsurface/plugins/_well_analysis/_views/_well_overview_view/_settings/_layout_options.py
@@ -38,6 +38,11 @@ class WellOverviewLayoutOptions(SettingsGroupABC):
                                     "label": "White background",
                                     "value": "white_background",
                                 },
+                                {
+                                    "label": "Error bars (P10/P90)",
+                                    "value": "errorbars",
+                                    "disabled": False,
+                                },
                             ],
                             value=["legend"],
                         ),